fix: send UDP errors to the correct GenServer
`Connector.udp_error` was `cast`ing an error message to the configured `reporter` pid, which was the `UDP.ErrorHandler`. However, the `handle_cast` implementation was in `UDP.Connector` instead, so any UDP errors were being sent to the wrong GenServer.

This PR registers the `UDP.Connector` GenServer with a name derived from `config.reporter_name` and then `cast`s the error message to that GenServer by name.

Co-authored-by: Delmer Reed <delmer814+1@gmail.com>