Allow client applications to specify event timestamps
By default, events are sent to InfluxDB without a timestamp. InfluxDB
will then assign a timestamp when it receives the event.

This change allows the client application to specify the event timestamp
by adding a `_timestamp` key to the event metadata. This is useful if an
application wants to emit historical events, or when the event handler
gets behind and a queue of events builds up, resulting in a delay in
sending them to InfluxDB.

The key name starts with a leading underscore to (hopefully) avoid
collisions with any metadata keys that client applications might already
be using.